refactor; fix linter warnings/errors
