+++ .eslintrc
@@ -40,7 +40,8 @@
         "@typescript-eslint/lines-between-class-members": ["error", "always",  { "exceptAfterSingleLine": true }],
         "@typescript-eslint/no-empty-function": ["warn", { "allow": ["methods", "constructors"] }],
         "@typescript-eslint/no-inferrable-types": "off",
-        "@typescript-eslint/no-var-requires": "off"
+        "@typescript-eslint/no-var-requires": "off",
+        "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" } ]
       }
     }
   ],
