+++ src/db/index.ts
@@ -1,0 +1,4 @@
+export * from './errors';
+export * from './games';
+export * from './player';
+export * from './rooms';
