+++ src/models/index.ts
@@ -1,0 +1,4 @@
+export * from './games.model';
+export * from './room.model';
+export * from './user.model';
+export * from './ws.model';
