--- src/ws/index.ts
@@ -1,1 +1,0 @@
-export * from './ws';
